--- python/ray/_private/job_manager/__init__.py
@@ -1,6 +1,0 @@
-from ray._private.job_manager.job_manager import (
-    JobManager, JobStatus, JobStatusStorageClient, JOB_ID_METADATA_KEY)
-
-__all__ = [
-    "JobManager", "JobStatus", "JobStatusStorageClient", "JOB_ID_METADATA_KEY"
-]
