+++ python/ray/tests/BUILD
@@ -306,14 +306,6 @@ py_test(
     deps = ["//:ray_lib"],
 )
 
-py_test(
-    name = "test_job_manager",
-    size = "small",
-    srcs = SRCS + ["test_job_manager.py"] + glob(["subprocess_driver_scripts/**/*.py"]),
-    tags = ["exclusive", "team:serve"],
-    deps = ["//:ray_lib"],
-)
-
 # TODO: use py_test(env = ...) in the build file with bazel 4.0
 py_test(
   name = "test_tracing",
